isogram: fix build issues
The test file was trying to import the example code which it shouldn't
do. Also removed unneeded package declarations and renamed the test
file to be consistent.
